+++ packages/custom_lint/example/example_lint/tools/custom_lint_debug.dart
@@ -1,5 +1,5 @@
 import 'dart:io';
-import 'package:custom_lint/basic_runner.dart';
+import 'package:custom_lint/custom_lint.dart';
 
 Future<void> main() async {
   await customLint(workingDirectory: Directory.current.parent);