+++ packages/custom_lint/test/run_plugin.dart
@@ -2,9 +2,10 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:analyzer_plugin/protocol/protocol_generated.dart';
-import 'package:custom_lint/src/analyzer_plugin/analyzer_plugin.dart';
-import 'package:custom_lint/src/analyzer_plugin/plugin_delegate.dart';
+import 'package:custom_lint/src/plugin_delegate.dart';
 import 'package:custom_lint/src/runner.dart';
+import 'package:custom_lint/src/server_isolate_channel.dart';
+import 'package:custom_lint/src/v2/custom_lint_analyzer_plugin.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
@@ -12,21 +13,24 @@ Future<List<AnalysisErrorsParams>> runServerInCliModeForApp(
   Directory directory,
   // to ignoreErrors as we cannot explictly handle errors
 ) async {
-  final runner = await startRunnerForApp(directory);
+  final runner = startRunnerForApp(directory);
   return runner.getLints(reload: false);
 }
 
-Future<CustomLintRunner> startRunnerForApp(
+CustomLintRunner startRunnerForApp(
   Directory directory, {
   bool ignoreErrors = false,
-}) async {
+}) {
+  final channel = ServerIsolateChannel();
   final runner = CustomLintRunner(
-    CustomLintPlugin(
+    // TODO use IO override to mock & test stdout/stderr
+    CustomLintServer.start(
+      sendPort: channel.receivePort.sendPort,
       delegate: CommandCustomLintDelegate(),
-      includeBuiltInLints: false,
       watchMode: false,
     ),
     directory,
+    channel,
   );
   addTearDown(runner.close);
 
@@ -40,7 +44,8 @@ Future<CustomLintRunner> startRunnerForApp(
       });
   }
 
-  await runner.initialize();
+  unawaited(runner.initialize);
+
   return runner;
 }